Revision complete assignment 40

+++ app/controllers/posts_controller.rb
@@ -51,14 +51,17 @@ class PostsController < ApplicationController
 
    def destroy
         @post = Post.find(params[:id])
-
-        if @post.destroy
-          flash[:notice] = "\"#{@post.title}\" was deleted successfully."
-          redirect_to @post.topic
-        else
-          flash.now[:alert] = "There was an error deleting the post."
-          render :show
+        #########
+        if current_user == @post.user || current_user.admin? then
+          if @post.destroy
+            flash[:notice] = "\"#{@post.title}\" was deleted successfully."
+            redirect_to @post.topic
+          else
+            flash.now[:alert] = "There was an error deleting the post."
+            render :show
+          end
         end
+      #####
       end
 
 #########
